Update README.md
For challenge 3, I used "for" loops because I had to incorporate the range() function in order to set the range parameters between 1 and 10 for the multiplication table. For challenge 1, I used "while" loops because it needs to continue executing the set of steps as long as the condition is true. I used Google Gemini to help me debug and explain to me why pieces of my code weren't working. I also used it to further explain to me how the range() function and the whitespaces worked in spacing out the multiplication table.